One more structure refactoring
