+++ Makefile
@@ -7,8 +7,17 @@ build:
 up-all:
 	docker compose up --build
 
+.PHONY: up-infra up-local
+
 up-infra:
-	docker compose up postgres migration --build
+	docker compose up -d postgres migration --build
+
+up-local: up-infra
+	poetry install
+	poetry run python -m aggregator
+
+down:
+	docker compose down
 
 render_diagrams:
 	./utilities/render_diagram.sh
