+++ alembic/env.py
@@ -2,8 +2,8 @@ from logging.config import fileConfig
 
 from sqlalchemy import engine_from_config, pool
 
+from aggregator.config import DB_CONNECTION_STRING
 from alembic import context
-from src.config import DB_CONNECTION_STRING
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -16,7 +16,7 @@ if config.config_file_name is not None:
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-from src import models
+from aggregator import models
 
 target_metadata = models.Base.metadata
 # target_metadata = None
