+++ pyproject.toml
@@ -4,7 +4,9 @@ version = "0.1.0"
 description = ""
 authors = ["Andrei Borisevich <andreyborisevichleti@gmail.com>"]
 readme = "README.md"
-packages = [{include = "src"}]
+packages = [
+    { include = "aggregator", from = "src" },
+  ]
 
 [tool.poetry.dependencies]
 python = "^3.12"
