+++ src/aggregator/__main__.py
@@ -2,11 +2,11 @@ from loguru import logger
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from bot.create_bot import create_bot
-from bot.warden.warden import Warden
-from config import DB_CONNECTION_STRING
-from posts_storage import PostStorage
-from telegram_agent.create_agent import create_telegram_agent
+from aggregator.bot.create_bot import create_bot
+from aggregator.bot.warden.warden import Warden
+from aggregator.config import DB_CONNECTION_STRING
+from aggregator.posts_storage import PostStorage
+from aggregator.telegram_agent.create_agent import create_telegram_agent
 
 if __name__ == "__main__":
     logger.info("Starting application...")
