+++ tests/test_warden.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 import pytest
 from pendulum import DateTime, Timezone
 
-from bot.warden.warden import (
+from aggregator.bot.warden.warden import (
     NotAllowed,
     Rule8to11EveryDay,
     Rule11to12Workdays,
